Renaming native_array2 to native_array (it was initially named that way to make the initial refactoring diff cleaner)
