+++ abra_core/src/builtins/native/mod.rs
@@ -3,7 +3,7 @@
 mod test_utils;
 
 mod common;
-mod native_array2;
+mod native_array;
 mod native_float;
 mod native_int;
 mod native_map;
@@ -12,7 +12,7 @@ mod native_string;
 
 pub use common::{NativeType, to_string, default_to_string_method};
 
-pub use native_array2::Array;
+pub use native_array::Array;
 pub use native_float::NativeFloat;
 pub use native_int::NativeInt;
 pub use native_map::NativeMap;
